Move MessageTransportException to a separate source file.
