+++ src/Host/Client/Impl/IMessageTransport.cs
@@ -7,22 +7,4 @@ namespace Microsoft.R.Host.Client {
         Task SendAsync(string message, CancellationToken ct = default(CancellationToken));
         Task<string> ReceiveAsync(CancellationToken ct = default(CancellationToken));
     }
-
-    [Serializable]
-    public class MessageTransportException : Exception {
-        public MessageTransportException() {
-        }
-
-        public MessageTransportException(string message)
-            : base(message) {
-        }
-
-        public MessageTransportException(string message, Exception innerException)
-            : base(message, innerException) {
-        }
-
-        public MessageTransportException(Exception innerException)
-            : this(innerException.Message, innerException) {
-        }
-    }
 }
