+++ src/Host/Client/Impl/MessageTransportException.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.R.Host.Client {
+    [Serializable]
+    public class MessageTransportException : Exception {
+        public MessageTransportException() {
+        }
+
+        public MessageTransportException(string message)
+            : base(message) {
+        }
+
+        public MessageTransportException(string message, Exception innerException)
+            : base(message, innerException) {
+        }
+
+        public MessageTransportException(Exception innerException)
+            : this(innerException.Message, innerException) {
+        }
+    }
+}
