+++ src/Host/Client/Impl/Microsoft.R.Host.Client.csproj
@@ -63,6 +63,7 @@
     <Compile Include="..\..\..\GlobalAssemblyInfo.cs">
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="MessageTransportException.cs" />
     <Compile Include="IMessageTransport.cs" />
     <Compile Include="IRContext.cs" />
     <Compile Include="IRCallbacks.cs" />
